Moved the Queues out of InternalStreamConnection
To allow for testing.
